chore: extract `divide` as a separate function
